chore/updates the behavior when we have some or no news
